Por fin terminada esta mierda... por el momento... espero no volver a ver esta rama xd
