+++ parse/ft_args3.c
@@ -1,0 +1,49 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_args3.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: emgarcia <emgarcia@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/11/05 21:12:27 by emgarcia          #+#    #+#             */
+/*   Updated: 2021/11/05 21:12:51 by emgarcia         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "parse.h"
+
+size_t	ft_contsp(t_general *g, size_t i)
+{
+	size_t	ncomnds;
+
+	ncomnds = 0;
+	while (i < g->parse.comndssize && ft_strlen(g->parse.comnds[i++]) == 1)
+		ncomnds++;
+	return (ncomnds);
+}
+
+void	ft_iniarg(t_general *g, size_t *j, char *str)
+{
+	g->args[*j].content = str;
+	if (ft_strlen(str) == 1)
+	{
+		if (str[0] == '<')
+			g->args[*j].type = 1;
+		if (str[0] == '>')
+			g->args[*j].type = 2;
+		if (str[0] == '|')
+			g->args[*j].type = 5;
+		if (str[0] == ';')
+			g->args[*j].type = 6;
+	}
+	else if (ft_strlen(str) == 2)
+	{
+		if (str[0] == '>')
+			g->args[*j].type = 7;
+		if (str[0] == '<')
+			g->args[*j].type = 8;
+	}
+	else
+		g->args[*j].type = 3;
+	*j += 1;
+}
