+++ parse/ft_aux2.c
@@ -6,7 +6,7 @@
 /*   By: emgarcia <emgarcia@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/28 04:11:14 by emgarcia          #+#    #+#             */
-/*   Updated: 2021/11/04 21:41:28 by emgarcia         ###   ########.fr       */
+/*   Updated: 2021/11/05 21:38:24 by emgarcia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,6 +32,17 @@ t_arg	*ft_copycleanargs(t_general *g)
 	}
 	free(g->args);
 	return (tmp);
+}
+
+void	ft_freecontent(t_arg *args)
+{
+	size_t	i;
+
+	i = -1;
+	printf("Llegue\n");
+	while (args[++i].content)
+		free(args[i].content);
+	free(args);
 }
 
 void	ft_freedouble(char **split)
